+++ trunk/source/ServerSource/liferay-plugins-sdk-6.1.1/portlets/DishStore-portlet/docroot/WEB-INF/service/irestads/service/DishServiceWrapper.java
@@ -57,7 +57,7 @@ public class DishServiceWrapper implements DishService,
 
 	public irestads.model.Dish createDish(java.lang.String dishName,
 		java.lang.String decription, java.lang.String avatarImg,
-		java.lang.String detailImg, java.lang.String detail, double referPrice,
+		java.lang.String detailImg, java.lang.String detail, int referPrice,
 		long categoryId) {
 		return _dishService.createDish(dishName, decription, avatarImg,
 			detailImg, detail, referPrice, categoryId);
